:hammer: Implement DialogService in commonMain for multi-platform reuse
